[REFACTOR] Extract setting queries and mutations on admin pages
