+++ src/client/react/admin/ServerSettingsView/index.js
@@ -1,39 +1,27 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import { autobind } from 'core-decorators';
-import { gql, graphql } from 'react-apollo';
+import { graphql } from 'react-apollo';
 import { Spinner, Button } from '@blueprintjs/core';
 import { connect } from 'react-redux';
 import { saveState } from '../../../actions/stateActions';
 import DateFormat from 'dateformat';
+import { getSettings } from '../../../graphql/setting';
 import Setting from './Setting';
 import RefreshBar from '../RefreshBar';
 import ViewError from '../ViewError';
 
 
-const QuerySettings = gql`
-query GetSettings($skip:Int,$limit:Int){
-	getSettings(skip:$skip,limit:$limit){
-		key
-		valueType
-		value
-		values
-		modified
-		modifiedBy
-	}
-}`;
+const QueryGetSettingsParams = 'key valueType value values modified modifiedBy';
 
-const QuerySettingsOptions = {
-	name: 'QuerySettings',
+const QueryGetSettingsOptions = {
+	name: 'QuerySettings', 
 	options: {
-			variables: {
-				skip: 0,
-				limit: 0
-		}
+		variables: { skip: 0, limit: 0 }
 	}
 }
 
-@graphql(QuerySettings, QuerySettingsOptions)
+@graphql(getSettings(QueryGetSettingsParams), QueryGetSettingsOptions)
 @connect()
 @autobind
 class ServerSettingsView extends React.Component {
